Moved helper function in a single file
