+++ src/models/model.ts
@@ -1,6 +1,6 @@
 import { isArray, isEqual, pick, omit } from 'lodash';
 import { plainToClass } from 'class-transformer';
-import { isUUID } from '../util/uuid';
+import { isUUID } from '../util/helpers';
 import wappsto from '../util/http_wrapper';
 import { printHttpError } from '../util/http_wrapper';
 import { printError } from '../util/debug';
