--- src/util/browser.ts
@@ -1,6 +1,0 @@
-export function isBrowser(): boolean {
-    return (
-        typeof window !== 'undefined' &&
-        Object.prototype.toString.call(window) === '[object Window]'
-    );
-}
