--- src/util/check_list.ts
@@ -1,8 +1,0 @@
-export function checkList(list: any[], check: any): boolean {
-    for (let i = 0; i < list.length; i++) {
-        if (list[i] === check || list[i].toString() === check.toString()) {
-            return true;
-        }
-    }
-    return false;
-}
